caps/tests/mochitest/test_loadLocalFont.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script type="application/javascript">

try {
  let LocalFiraSans = new FontFace(
    "LocalFiraSans",
    "url('chrome://pocket/content/panels/fonts/FiraSans-Regular.woff') format('woff')"
  ).load().then(() => {
    ok(true, "We should suuceed");
  }, () => {
    ok(false, "We shouldn't fail (reject)");
  });
  console.log(LocalFiraSans); // avoid warning
} catch (e) {
  ok(false, "We shouldn't fail (catch)" + e);
}

</script>
</pre>
</body>
</html>
